+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Avro schema form file avsc
+
+### Fixed
+
+- Upgrade kafka client version
+- Upgrade textual version
+
 ## [3.0.0] - 2024-11-24
 
 ### Added